geocode/requirements.txt
requests
aiohttp
shapely
pyproj
python-dotenv
